+++ packages/quantic/cypress/integration/case-classification/case-classification.cypress.ts
@@ -82,11 +82,8 @@ describe('quantic-case-classification', () => {
         Expect.numberOfSuggestions(suggestionsCount);
         Expect.correctSugestionsOrder(allOptions.slice(0, suggestionsCount));
         Expect.numberOfInlineOptions(0);
-        Expect.logClickedSuggestion(firstSuggestionIndex);
-        Expect.logUpdatedClassificationFromSuggestion(
-          coveoDefaultField,
-          firstSuggestionIndex
-        );
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
       });
 
       scope('when selecting a suggestion', () => {
@@ -98,6 +95,7 @@ describe('quantic-case-classification', () => {
           coveoDefaultField,
           clickedIndex
         );
+        Expect.correctValue(allOptions[clickedIndex].value);
       });
 
       scope('when selecting an option from the select input', () => {
@@ -111,6 +109,7 @@ describe('quantic-case-classification', () => {
           coveoDefaultField,
           clickedIndex
         );
+        Expect.correctValue(allOptions[clickedIndex].value);
       });
     });
   });
@@ -198,11 +197,8 @@ describe('quantic-case-classification', () => {
         Expect.numberOfInlineOptions(0);
         Expect.displaySelectInput(false);
         Expect.displaySelectTitle(true);
-        Expect.logClickedSuggestion(firstSuggestionIndex);
-        Expect.logUpdatedClassificationFromSuggestion(
-          coveoDefaultField,
-          firstSuggestionIndex
-        );
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
       });
 
       scope('when selecting an option from the select input', () => {
@@ -215,6 +211,7 @@ describe('quantic-case-classification', () => {
           coveoDefaultField,
           clickedIndex
         );
+        Expect.correctValue(allOptions[clickedIndex].value);
       });
     });
   });
@@ -255,6 +252,7 @@ describe('quantic-case-classification', () => {
           coveoDefaultField,
           clickedIndex
         );
+        Expect.correctValue(allOptions[clickedIndex].value);
       });
     });
 
@@ -293,6 +291,7 @@ describe('quantic-case-classification', () => {
 
   describe('when maxSuggestions is set to the number of options', () => {
     const optionsCount = 5;
+    const suggestionsCount = 2;
     it('should display all the options as inline options', () => {
       visitCaseClassification({
         maxSuggestions: optionsCount,
@@ -307,7 +306,6 @@ describe('quantic-case-classification', () => {
       });
 
       scope('when fetching suggestions', () => {
-        const suggestionsCount = 2;
         const firstSuggestionIndex = 0;
 
         mockCaseClassification(
@@ -319,11 +317,8 @@ describe('quantic-case-classification', () => {
         Expect.numberOfSuggestions(suggestionsCount);
         Expect.numberOfInlineOptions(allOptions.length - suggestionsCount);
         Expect.displaySelectInput(false);
-        Expect.logClickedSuggestion(firstSuggestionIndex);
-        Expect.logUpdatedClassificationFromSuggestion(
-          coveoDefaultField,
-          firstSuggestionIndex
-        );
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
       });
 
       scope('when selecting a suggestion', () => {
@@ -335,6 +330,7 @@ describe('quantic-case-classification', () => {
           coveoDefaultField,
           clickedIndex
         );
+        Expect.correctValue(allOptions[clickedIndex].value);
       });
 
       scope('when selecting an inline option', () => {
@@ -345,6 +341,7 @@ describe('quantic-case-classification', () => {
           coveoDefaultField,
           clickedIndex
         );
+        Expect.correctValue(allOptions[clickedIndex + suggestionsCount].value);
       });
     });
   });
@@ -377,11 +374,8 @@ describe('quantic-case-classification', () => {
         Expect.numberOfSuggestions(suggestionsCount);
         Expect.numberOfInlineOptions(allOptions.length - suggestionsCount);
         Expect.displaySelectInput(false);
-        Expect.logClickedSuggestion(firstSuggestionIndex);
-        Expect.logUpdatedClassificationFromSuggestion(
-          coveoDefaultField,
-          firstSuggestionIndex
-        );
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
       });
     });
   });
@@ -433,13 +427,11 @@ describe('quantic-case-classification', () => {
         Expect.displaySelectTitle(true);
         Expect.displaySelectInput(false);
         Expect.numberOfSuggestions(suggestionsCount);
-        Expect.logClickedSuggestion(firstSuggestionIndex);
-        Expect.logUpdatedClassificationFromSuggestion(
-          coveoDefaultField,
-          firstSuggestionIndex
-        );
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
         Actions.clickSuggestion(firstSuggestionIndex);
         Expect.logDeselect(coveoDefaultField);
+        Expect.correctValue('');
         Actions.reportValidity();
         Expect.displayError(true);
       });
@@ -456,6 +448,7 @@ describe('quantic-case-classification', () => {
         Expect.displaySelectInput(false);
         Expect.numberOfSuggestions(suggestionsCount);
         Actions.clickSuggestion(firstSuggestionIndex);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
         Expect.logClickedSuggestion(firstSuggestionIndex);
         Expect.logUpdatedClassificationFromSuggestion(
           coveoDefaultField,
@@ -463,6 +456,64 @@ describe('quantic-case-classification', () => {
         );
         Actions.reportValidity();
         Expect.displayError(false);
+      });
+    });
+
+    it('should hide the error after fetching suggestions and autoselecting the one with the highest confidence', () => {
+      visitCaseClassification({
+        required: true,
+      });
+
+      scope('when reporting validity and no option is selected', () => {
+        Expect.displayLabel(true);
+        Expect.numberOfInlineOptions(0);
+        Expect.numberOfSuggestions(0);
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(true);
+        Actions.reportValidity();
+        Expect.displayError(true);
+      });
+
+      scope('when fetching suggestions', () => {
+        const suggestionsCount = 2;
+        const firstSuggestionIndex = 0;
+
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(0, suggestionsCount)
+        );
+        fetchClassifications();
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
+        Expect.displaySelectTitle(true);
+        Expect.displaySelectInput(false);
+        Expect.numberOfSuggestions(suggestionsCount);
+        Expect.displayError(false);
+      });
+    });
+
+    it('should keep the error after fetching suggestions but not finding any', () => {
+      visitCaseClassification({
+        required: true,
+      });
+
+      scope('when reporting validity and no suggestion is selected', () => {
+        Expect.displayLabel(true);
+        Expect.numberOfInlineOptions(0);
+        Expect.numberOfSuggestions(0);
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(true);
+        Actions.reportValidity();
+        Expect.displayError(true);
+      });
+
+      scope('when fetching suggestions', () => {
+        mockCaseClassification(coveoDefaultField, []);
+        fetchClassifications();
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(true);
+        Expect.numberOfSuggestions(0);
+        Expect.displayError(true);
       });
     });
   });
@@ -501,6 +552,272 @@ describe('quantic-case-classification', () => {
         Expect.displaySelectTitle(false);
         Expect.displaySelectInput(false);
         Expect.displayRenderingError(true, missingCoveoFieldNameError);
+      });
+    });
+  });
+
+  describe('when selecting a suggestion and then receiving new suggestions', () => {
+    it('should keep the suggestion selected by the user', () => {
+      const clickedIndex = 1;
+      visitCaseClassification({
+        maxSuggestions: defaultMaxSuggestions,
+      });
+
+      scope('when loading the page', () => {
+        Expect.displayLabel(true);
+        Expect.numberOfInlineOptions(0);
+        Expect.numberOfSuggestions(0);
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(true);
+      });
+
+      scope('when fetching suggestions', () => {
+        const suggestionsCount = 3;
+        const firstSuggestionIndex = 0;
+
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(0, suggestionsCount)
+        );
+        fetchClassifications();
+        Expect.displaySelectTitle(true);
+        Expect.displaySelectInput(false);
+        Expect.numberOfSuggestions(suggestionsCount);
+        Expect.correctSugestionsOrder(allOptions.slice(0, suggestionsCount));
+        Expect.numberOfInlineOptions(0);
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
+      });
+
+      scope('when selecting a suggestion', () => {
+        Actions.clickSuggestion(clickedIndex);
+        Expect.logClickedSuggestion(clickedIndex);
+        Expect.logUpdatedClassificationFromSuggestion(
+          coveoDefaultField,
+          clickedIndex
+        );
+        Expect.correctValue(allOptions[clickedIndex].value);
+      });
+
+      scope('when fetching suggestions again', () => {
+        const suggestionsCount = 3;
+
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(0, suggestionsCount)
+        );
+        fetchClassifications();
+        Expect.displaySelectTitle(true);
+        Expect.displaySelectInput(false);
+        Expect.numberOfSuggestions(suggestionsCount);
+        Expect.correctSugestionsOrder(allOptions.slice(0, suggestionsCount));
+        Expect.numberOfInlineOptions(0);
+        Expect.correctValue(allOptions[clickedIndex].value);
+      });
+    });
+  });
+
+  describe('when selecting a specific suggestion and then receiving new suggestions that does not contain the previously selected option', () => {
+    it('should keep the suggestion selected by the user and display it in the select input', () => {
+      const clickedIndex = 2;
+      visitCaseClassification({
+        maxSuggestions: defaultMaxSuggestions,
+      });
+
+      scope('when loading the page', () => {
+        Expect.displayLabel(true);
+        Expect.numberOfInlineOptions(0);
+        Expect.numberOfSuggestions(0);
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(true);
+      });
+
+      scope('when fetching suggestions', () => {
+        const suggestionsCount = 3;
+        const firstSuggestionIndex = 0;
+
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(0, suggestionsCount)
+        );
+        fetchClassifications();
+        Expect.displaySelectTitle(true);
+        Expect.displaySelectInput(false);
+        Expect.numberOfSuggestions(suggestionsCount);
+        Expect.correctSugestionsOrder(allOptions.slice(0, suggestionsCount));
+        Expect.numberOfInlineOptions(0);
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
+      });
+
+      scope('when selecting a suggestion', () => {
+        Actions.clickSuggestion(clickedIndex);
+        Expect.logClickedSuggestion(clickedIndex);
+        Expect.logUpdatedClassificationFromSuggestion(
+          coveoDefaultField,
+          clickedIndex
+        );
+        Expect.correctValue(allOptions[clickedIndex].value);
+      });
+
+      scope('when fetching suggestions again', () => {
+        const suggestionsCount = 1;
+
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(0, suggestionsCount)
+        );
+        fetchClassifications();
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(false);
+        Expect.numberOfSuggestions(0);
+        Expect.numberOfInlineOptions(0);
+        Expect.correctValue(allOptions[clickedIndex].value);
+      });
+    });
+  });
+
+  describe('when selecting an option from the select input and then fetching suggestions', () => {
+    it('should keep the option selected by the user, display it in the select input and hide the suggestions', () => {
+      const clickedIndex = 3;
+      visitCaseClassification({
+        maxSuggestions: defaultMaxSuggestions,
+      });
+
+      scope('when loading the page', () => {
+        Expect.displayLabel(true);
+        Expect.numberOfInlineOptions(0);
+        Expect.numberOfSuggestions(0);
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(true);
+      });
+
+      scope('when selecting an option from the select input', () => {
+        Actions.openSelectInput();
+        Actions.clickSelectOption(clickedIndex);
+        Expect.logUpdatedClassificationFromSelectOption(
+          coveoDefaultField,
+          clickedIndex
+        );
+        Expect.correctValue(allOptions[clickedIndex].value);
+      });
+
+      scope('when fetching suggestions', () => {
+        const suggestionsCount = 3;
+
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(0, suggestionsCount)
+        );
+        fetchClassifications();
+        Expect.displayLoading(false);
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(true);
+        Expect.numberOfSuggestions(0);
+        Expect.numberOfInlineOptions(0);
+        Expect.correctValue(allOptions[clickedIndex].value);
+      });
+    });
+  });
+
+  describe('when receiving new suggestions without changing the by default auto-selected suggestion', () => {
+    it('should auto-select the suggestion with the highest confidence from the newly recieved suggestions', () => {
+      visitCaseClassification({
+        maxSuggestions: defaultMaxSuggestions,
+      });
+
+      scope('when loading the page', () => {
+        Expect.displayLabel(true);
+        Expect.numberOfInlineOptions(0);
+        Expect.numberOfSuggestions(0);
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(true);
+      });
+
+      scope('when fetching suggestions', () => {
+        const suggestionsCount = 3;
+        const firstSuggestionIndex = 0;
+
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(0, suggestionsCount)
+        );
+        fetchClassifications();
+        Expect.displaySelectTitle(true);
+        Expect.displaySelectInput(false);
+        Expect.numberOfSuggestions(suggestionsCount);
+        Expect.correctSugestionsOrder(allOptions.slice(0, suggestionsCount));
+        Expect.numberOfInlineOptions(0);
+        Expect.logClickedSuggestion(firstSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
+      });
+
+      scope('when fetching suggestions again', () => {
+        const suggestionsCount = 2;
+        const firstDisplayedSuggestionIndex = 0;
+        const firstSuggestionIndex = 1;
+
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(
+            firstSuggestionIndex,
+            firstSuggestionIndex + suggestionsCount
+          )
+        );
+        fetchClassifications();
+        Expect.displaySelectTitle(true);
+        Expect.displaySelectInput(false);
+        Expect.numberOfSuggestions(suggestionsCount);
+        Expect.correctSugestionsOrder(
+          allOptions.slice(
+            firstSuggestionIndex,
+            firstSuggestionIndex + suggestionsCount
+          )
+        );
+        Expect.numberOfInlineOptions(0);
+        Expect.logClickedSuggestion(firstDisplayedSuggestionIndex, true);
+        Expect.correctValue(allOptions[firstSuggestionIndex].value);
+      });
+    });
+  });
+
+  describe('when selecting an inline option and then fetching suggestions', () => {
+    it('should keep the option selected by the user, display it in an inline option and display the suggestions', () => {
+      const suggestionsCount = 3;
+      const clickedIndex = 4;
+      visitCaseClassification({
+        maxSuggestions: allOptions.length,
+      });
+
+      scope('when loading the page', () => {
+        Expect.displayLabel(true);
+        Expect.numberOfInlineOptions(allOptions.length);
+        Expect.numberOfSuggestions(0);
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(false);
+      });
+
+      scope('when selecting an inline option', () => {
+        Actions.clickInlineOption(clickedIndex);
+        Expect.logUpdatedClassificationFromInlineOption(
+          coveoDefaultField,
+          clickedIndex
+        );
+        Expect.correctValue(allOptions[clickedIndex].value);
+      });
+
+      scope('when fetching suggestions', () => {
+        mockCaseClassification(
+          coveoDefaultField,
+          allOptions.slice(0, suggestionsCount)
+        );
+        fetchClassifications();
+        Expect.displaySelectTitle(false);
+        Expect.displaySelectInput(false);
+        Expect.numberOfSuggestions(suggestionsCount);
+        Expect.correctSugestionsOrder(allOptions.slice(0, suggestionsCount));
+        Expect.numberOfInlineOptions(allOptions.length - suggestionsCount);
+        Expect.correctValue(allOptions[clickedIndex].value);
       });
     });
   });